+++ app/Stubs/Migrations/HasPaymentStats.php
@@ -14,9 +14,9 @@ trait HasPaymentStats
 {
     public function paymentStats(Blueprint $table): Blueprint
     {
-        $table->unsignedBigInteger('number_payment_records')->default(0);
-        $table->unsignedBigInteger('number_payments')->default(0);
-        $table->unsignedBigInteger('number_refunds')->default(0);
+        $table->unsignedInteger('number_payment_records')->default(0);
+        $table->unsignedInteger('number_payments')->default(0);
+        $table->unsignedInteger('number_refunds')->default(0);
 
 
         if ($table->getTable()=='shop_accounting_stats') {
@@ -35,9 +35,9 @@ trait HasPaymentStats
 
 
         foreach (PaymentStateEnum::cases() as $state) {
-            $table->unsignedBigInteger("number_payment_records_state_{$state->snake()}")->default(0);
-            $table->unsignedBigInteger("number_payments_state_{$state->snake()}")->default(0);
-            $table->unsignedBigInteger("number_refunds_state_{$state->snake()}")->default(0);
+            $table->unsignedInteger("number_payment_records_state_{$state->snake()}")->default(0);
+            $table->unsignedInteger("number_payments_state_{$state->snake()}")->default(0);
+            $table->unsignedInteger("number_refunds_state_{$state->snake()}")->default(0);
         }
 
         return $table;
