+++ database/migrations/2022_08_14_124137_create_timezones_table.php
@@ -16,7 +16,7 @@ return new class () extends Migration {
         Schema::create('timezones', function (Blueprint $table) {
             $table->smallIncrements('id');
             $table->string('name')->unique()->collation('und_ns_ci');
-            $table->bigInteger('offset')->nullable()->comment('in hours');
+            $table->integer('offset')->nullable()->comment('in hours');
             $table->unsignedSmallInteger('country_id')->nullable();
             $table->foreign('country_id')->references('id')->on('countries');
             $table->float('latitude')->nullable();
