+++ database/migrations/2022_08_26_170739_create_media_table.php
@@ -31,7 +31,7 @@ return new class () extends Migration {
             $table->string('mime_type')->nullable();
             $table->string('disk');
             $table->string('conversions_disk')->nullable();
-            $table->unsignedBigInteger('size');
+            $table->unsignedInteger('size');
             $table->json('manipulations');
             $table->json('custom_properties');
             $table->json('generated_conversions');
