+++ database/migrations/2023_04_10_075024_media_customer.php
@@ -16,7 +16,7 @@ return new class () extends Migration {
             $table->unsignedInteger('customer_id')->index();
             $table->string('type')->index();
             $table->foreign('customer_id')->references('id')->on('customers');
-            $table->unsignedBigInteger('media_id')->index();
+            $table->unsignedInteger('media_id')->index();
             $table->unique(['customer_id', 'media_id']);
             $table->string('owner_type')->index();
             $table->unsignedInteger('owner_id');
