+++ database/migrations/2023_04_10_081933_media_guest.php
@@ -16,7 +16,7 @@ return new class () extends Migration {
             $table->unsignedInteger('guest_id')->index();
             $table->string('type')->index();
             $table->foreign('guest_id')->references('id')->on('guests');
-            $table->unsignedBigInteger('media_id')->index();
+            $table->unsignedInteger('media_id')->index();
             $table->unique(['guest_id', 'media_id']);
             $table->string('owner_type')->index();
             $table->unsignedInteger('owner_id');
